fix: rename `strings.h` into `string_helper.h`
`strings` is apparently a possible include on some systems.
