+++ src/utils/string_helper.h
